# server/http2/openapi_generated_api.go
// Package http2 provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.2 DO NOT EDIT.
package http2

import (
	"context"
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get movie by id
	// (GET /movies/{movieId})
	GetMovie(w http.ResponseWriter, r *http.Request, movieId int64)
	// Get latest episodes
	// (GET /shows/latest-episodes)
	GetLatestEpisodes(w http.ResponseWriter, r *http.Request, params GetLatestEpisodesParams)
	// Get latest movies
	// (GET /shows/latest-movies)
	GetLatestMovies(w http.ResponseWriter, r *http.Request, params GetLatestMoviesParams)
	// Get latest tv series
	// (GET /shows/latest-tv-series)
	GetLatestTvSeries(w http.ResponseWriter, r *http.Request, params GetLatestTvSeriesParams)
	// Search shows
	// (GET /shows/search)
	SearchShows(w http.ResponseWriter, r *http.Request, params SearchShowsParams)
	// Get tv series by id
	// (GET /tv-series/{tvSeriesId})
	GetTvSeries(w http.ResponseWriter, r *http.Request, tvSeriesId int64)
	// Get tv season by tv series id and season number
	// (GET /tv-series/{tvSeriesId}/seasons/{seasonNumber})
	GetTvSeason(w http.ResponseWriter, r *http.Request, tvSeriesId int64, seasonNumber int)
	// Get version
	// (GET /version)
	GetVersion(w http.ResponseWriter, r *http.Request)
	// Get video by id
	// (GET /videos/{id})
	GetVideo(w http.ResponseWriter, r *http.Request, id int64)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get movie by id
// (GET /movies/{movieId})
func (_ Unimplemented) GetMovie(w http.ResponseWriter, r *http.Request, movieId int64) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get latest episodes
// (GET /shows/latest-episodes)
func (_ Unimplemented) GetLatestEpisodes(w http.ResponseWriter, r *http.Request, params GetLatestEpisodesParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get latest movies
// (GET /shows/latest-movies)
func (_ Unimplemented) GetLatestMovies(w http.ResponseWriter, r *http.Request, params GetLatestMoviesParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get latest tv series
// (GET /shows/latest-tv-series)
func (_ Unimplemented) GetLatestTvSeries(w http.ResponseWriter, r *http.Request, params GetLatestTvSeriesParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Search shows
// (GET /shows/search)
func (_ Unimplemented) SearchShows(w http.ResponseWriter, r *http.Request, params SearchShowsParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get tv series by id
// (GET /tv-series/{tvSeriesId})
func (_ Unimplemented) GetTvSeries(w http.ResponseWriter, r *http.Request, tvSeriesId int64) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get tv season by tv series id and season number
// (GET /tv-series/{tvSeriesId}/seasons/{seasonNumber})
func (_ Unimplemented) GetTvSeason(w http.ResponseWriter, r *http.Request, tvSeriesId int64, seasonNumber int) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get version
// (GET /version)
func (_ Unimplemented) GetVersion(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get video by id
// (GET /videos/{id})
func (_ Unimplemented) GetVideo(w http.ResponseWriter, r *http.Request, id int64) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetMovie operation middleware
func (siw *ServerInterfaceWrapper) GetMovie(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "movieId" -------------
	var movieId int64

	err = runtime.BindStyledParameterWithLocation("simple", false, "movieId", runtime.ParamLocationPath, chi.URLParam(r, "movieId"), &movieId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "movieId", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMovie(w, r, movieId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetLatestEpisodes operation middleware
func (siw *ServerInterfaceWrapper) GetLatestEpisodes(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetLatestEpisodesParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "pageSize", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetLatestEpisodes(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetLatestMovies operation middleware
func (siw *ServerInterfaceWrapper) GetLatestMovies(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetLatestMoviesParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "pageSize", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetLatestMovies(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetLatestTvSeries operation middleware
func (siw *ServerInterfaceWrapper) GetLatestTvSeries(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params GetLatestTvSeriesParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "pageSize", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetLatestTvSeries(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// SearchShows operation middleware
func (siw *ServerInterfaceWrapper) SearchShows(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	// Parameter object where we will unmarshal all parameters from the context
	var params SearchShowsParams

	// ------------- Required query parameter "query" -------------

	if paramValue := r.URL.Query().Get("query"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "query"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "query", r.URL.Query(), &params.Query)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "query", Err: err})
		return
	}

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "pageSize", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.SearchShows(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetTvSeries operation middleware
func (siw *ServerInterfaceWrapper) GetTvSeries(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "tvSeriesId" -------------
	var tvSeriesId int64

	err = runtime.BindStyledParameterWithLocation("simple", false, "tvSeriesId", runtime.ParamLocationPath, chi.URLParam(r, "tvSeriesId"), &tvSeriesId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "tvSeriesId", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTvSeries(w, r, tvSeriesId)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetTvSeason operation middleware
func (siw *ServerInterfaceWrapper) GetTvSeason(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "tvSeriesId" -------------
	var tvSeriesId int64

	err = runtime.BindStyledParameterWithLocation("simple", false, "tvSeriesId", runtime.ParamLocationPath, chi.URLParam(r, "tvSeriesId"), &tvSeriesId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "tvSeriesId", Err: err})
		return
	}

	// ------------- Path parameter "seasonNumber" -------------
	var seasonNumber int

	err = runtime.BindStyledParameterWithLocation("simple", false, "seasonNumber", runtime.ParamLocationPath, chi.URLParam(r, "seasonNumber"), &seasonNumber)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "seasonNumber", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTvSeason(w, r, tvSeriesId, seasonNumber)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetVersion operation middleware
func (siw *ServerInterfaceWrapper) GetVersion(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetVersion(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

// GetVideo operation middleware
func (siw *ServerInterfaceWrapper) GetVideo(w http.ResponseWriter, r *http.Request) {
	ctx := r.Context()

	var err error

	// ------------- Path parameter "id" -------------
	var id int64

	err = runtime.BindStyledParameterWithLocation("simple", false, "id", runtime.ParamLocationPath, chi.URLParam(r, "id"), &id)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx = context.WithValue(ctx, BearerAuthScopes, []string{})

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetVideo(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r.WithContext(ctx))
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/movies/{movieId}", wrapper.GetMovie)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/shows/latest-episodes", wrapper.GetLatestEpisodes)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/shows/latest-movies", wrapper.GetLatestMovies)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/shows/latest-tv-series", wrapper.GetLatestTvSeries)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/shows/search", wrapper.SearchShows)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/tv-series/{tvSeriesId}", wrapper.GetTvSeries)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/tv-series/{tvSeriesId}/seasons/{seasonNumber}", wrapper.GetTvSeason)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/version", wrapper.GetVersion)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/videos/{id}", wrapper.GetVideo)
	})

	return r
}